refactor: ScheduleDateResolver 리팩토링 #15 패키지 이동 null 반환 경우 삭제
